[Modify]log_in、log_outメソッドに処理を省略するガード節を追加
既にログインあるいはログアウトしていた場合にlog_in、log_outを
呼び出してDBアクセスが発生するのを防ぐため。

+++ app/helpers/sessions_helper.rb
@@ -1,10 +1,14 @@
 module SessionsHelper
   def log_in(user)
+    return if logged_in?
+
     session[:user_id] = user.id
     remember(user)
   end
 
   def log_out
+    return unless logged_in?
+
     forget(current_user)
     session.delete(:user_id)
     @current_user = nil
